FRD/requirements.txt
h5py==3.1.0
imageio==2.9.0
Markdown==3.3.4
mat73==0.58
matplotlib==3.3.4
numpy==1.19.2
opencv-python==4.5.3.56
panda==0.3.1
pandas==1.1.5
Pillow==8.3.1
pydensecrf==1.0rc3
pyparsing==2.4.7
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
timm==0.4.12
torch==1.7.1
torchvision==0.8.2
tqdm==4.62.2
